refactor project and add node-nats-straming

+++ auth/src/routes/admin-delete-users.ts
@@ -1,6 +1,7 @@
 import express , { Request , Response } from 'express';
 import { User } from '../models/user.model';
 import { requireAuth , BadRequestError } from '@social-microservices/common';
+import mongoose from 'mongoose';
 
 const router = express.Router();
 
@@ -17,6 +18,11 @@ router.delete('/api/auth/admin' , requireAuth , async(req : Request , res : Resp
       {
             throw new BadRequestError('User have no this permission');
       };
+
+      if (!req.query.id || !mongoose.Types.ObjectId.isValid(String(req.query.id))) 
+      {
+            throw new BadRequestError("Id Is Invalid");
+      }
 
       await User.findByIdAndRemove(req.query.id);
